net: Move update_nad_references to libs/net/vmspec
Move update_nad_references from tests/network/libs/nad_ref.py
into libs/net/vmspec.py alongside the VMI condition helpers it
depends on. Update callers in l2_bridge and localnet nad_ref_change
test suites accordingly.
