+++ tests/network/l2_bridge/nad_ref_change/test_nad_ref_change.py
@@ -14,7 +14,7 @@ Preconditions:
 import pytest
 
 from libs.net.ip import filter_link_local_addresses
-from libs.net.vmspec import lookup_iface_status
+from libs.net.vmspec import lookup_iface_status, update_nad_references
 from tests.network.l2_bridge.libl2bridge import LINUX_BRIDGE_IFACE_NAME_1, LINUX_BRIDGE_IFACE_NAME_2
 from tests.network.l2_bridge.nad_ref_change.lib_helpers import (
     GUEST_IFACE_1,
@@ -22,7 +22,6 @@ from tests.network.l2_bridge.nad_ref_change.lib_helpers import (
     assert_connectivity,
     assert_no_connectivity,
 )
-from tests.network.libs.nad_ref import update_nad_references
 
 
 @pytest.mark.usefixtures("baseline_connectivity")
